starters/ts-project-starter/ts-project/requirements.txt
torch
numpy
pandas
pyyaml
tqdm
